Code/requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==20.1.0
art==5.2
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==3.3.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
cycler==0.10.0
dataclasses==0.8
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
entrypoints==0.3
et-xmlfile==1.1.0
facenet-pytorch==2.5.2
filelock==3.0.12
Flask==1.1.2
Flask-Mail==0.9.1
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
hmmlearn==0.2.6
idna==3.2
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==4.0.0
importlib-resources==5.2.2
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mysql-connector==2.2.9
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.1
onetimepad==1.4
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.2.0
plotly==5.1.0
pluggy==0.13.1
progress==1.6
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.3
py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pydeck==0.6.2
PyEmotion==1.0.0
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
qrcode==7.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
streamlit==1.0.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
virtualenv==20.7.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
xlrd==2.0.1
xlsx2html==0.4.0
zipp==3.4.1
